pyproject.toml
[tool.poetry]
name = "house-plant-hub-api"
package-mode = false
version = "0.1.0"
description = ""
authors = ["Glen Sharp <glen.sharp@hotmail.co.uk>"]
license = "proprietary"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.1.13"
djangorestframework = "3.12.4"
django-cors-headers = "4.5.0"
django-extensions = "3.2.3"
PyJWT = "2.4.0"
pytz = "2021.1"
flake8 = "7.0.0"


[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.8.4"]
build-backend = "poetry.core.masonry.api"
